+++ src/build_novelyst_aeon2.py
@@ -7,7 +7,7 @@ The PyWriter project (see see https://github.com/peter88213/PyWriter)
 must be located on the same directory level as the aeon2yw project. 
 
 For further information see https://github.com/peter88213/novelyst_aeon2
-Published under the MIT License (https://opensource.org/licenses/mit-license.php)
+License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 import os
 import sys