+++ src/translate_de.py
@@ -31,7 +31,7 @@ File structure:
     
 Copyright (c) 2022 Peter Triesberger
 For further information see https://github.com/peter88213/novelyst_aeon2
-Published under the MIT License (https://opensource.org/licenses/mit-license.php)
+License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 import os
 import sys