+++ docs/changelog.md
@@ -6,7 +6,7 @@
 
 ### Planned features
 
-See the [GitHub "features" project](https://github.com/peter88213/aeon2yw_novelyst/projects/1).
+See the [GitHub "features" project](https://github.com/peter88213/novelyst_aeon2/projects/1).
 
 ### v0.4.0 Beta test release