+++ docs/changelog.md
@@ -8,6 +8,14 @@
 
 See the [GitHub "features" project](https://github.com/peter88213/novelyst_aeon2/projects/1).
 
+### v0.6.0
+
+- Release under the GPLv3 license.
+
+Compatibility: novelyst v1.0.0 API
+Based on PyWriter v7.4.9
+Based on aeon2yw v1.8.2
+
 ### v0.4.0 Beta test release
 
 - Provide information for the novelyst v0.50+ Plugin Manager.